# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Badminton Court Booking API",
        "description": "API untuk sistem pemesanan lapangan badminton",
        "contact": {
            "name": "Badminton Booking Support",
            "email": "admin@badminton-booking.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Development Server"
        }
    ],
    "paths": {
...
        "/api/v1/bookings": {
            "get": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Dapatkan semua booking",
                "description": "Mengambil semua data booking lapangan badminton beserta informasi jadwal dan lapangan",
                "operationId": "db78b9a82d12b0084666ec17d98def14",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data booking",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Bookings retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Booking"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Buat booking baru",
                "description": "Membuat booking baru untuk lapangan badminton. Sistem akan otomatis menghitung total harga berdasarkan durasi dan harga per jam lapangan.",
                "operationId": "b5c7215d7874daad77109bf4b8691e32",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "schedule_id",
                                    "customer_name",
                                    "customer_phone"
                                ],
                                "properties": {
                                    "schedule_id": {
                                        "description": "ID jadwal yang akan dibooking",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "customer_name": {
                                        "description": "Nama customer",
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "customer_phone": {
                                        "description": "Nomor telepon customer",
                                        "type": "string",
                                        "example": "08123456789"
                                    },
                                    "customer_email": {
                                        "description": "Email customer (opsional)",
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@email.com"
                                    },
                                    "notes": {
                                        "description": "Catatan tambahan (opsional)",
                                        "type": "string",
                                        "example": "Booking untuk latihan"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Booking berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Booking created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Booking"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error atau jadwal tidak tersedia",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/bookings/{id}": {
            "get": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Dapatkan detail booking",
                "description": "Mengambil detail booking berdasarkan ID beserta informasi jadwal dan lapangan",
                "operationId": "e0de9c8b26a872ae857b77351497c4dc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID booking",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil detail booking",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Booking retrieved successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Booking"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Booking tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Update booking",
                "description": "Mengupdate data booking (data customer dan catatan saja, tidak bisa mengubah jadwal)",
                "operationId": "fa5137829b9ea38ece5bfc8b43060427",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID booking",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "customer_name": {
                                        "description": "Nama customer",
                                        "type": "string",
                                        "example": "John Doe Updated"
                                    },
                                    "customer_phone": {
                                        "description": "Nomor telepon customer",
                                        "type": "string",
                                        "example": "08123456790"
                                    },
                                    "customer_email": {
                                        "description": "Email customer",
                                        "type": "string",
                                        "format": "email",
                                        "example": "john.updated@email.com"
                                    },
                                    "notes": {
                                        "description": "Catatan tambahan",
                                        "type": "string",
                                        "example": "Catatan yang diupdate"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Booking berhasil diupdate",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Booking updated successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Booking"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Booking tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Hapus booking",
                "description": "Menghapus booking dan mengubah status jadwal kembali menjadi 'available'",
                "operationId": "ec2dd19088b1d73ed00d0d42be1d694e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID booking",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Booking berhasil dihapus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Booking deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Booking tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Schedule": {
                "title": "Schedule",
                "description": "Model jadwal lapangan",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "court_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-01-15"
                    },
                    "start_time": {
                        "type": "string",
                        "format": "time",
                        "example": "09:00"
                    },
                    "end_time": {
                        "type": "string",
                        "format": "time",
                        "example": "11:00"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "available",
                            "booked"
                        ],
                        "example": "available"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15T10:00:00Z"
                    },
                    "court": {
                        "$ref": "#/components/schemas/Court"
                    }
                },
                "type": "object"
            },
            "Booking": {
                "title": "Booking",
                "description": "Model pemesanan lapangan",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "schedule_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "customer_name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "customer_phone": {
                        "type": "string",
                        "example": "08123456789"
                    },
                    "customer_email": {
                        "type": "string",
                        "example": "john@email.com"
                    },
                    "total_price": {
                        "type": "number",
                        "format": "decimal",
                        "example": 100000
                    },
                    "notes": {
                        "type": "string",
                        "example": "Booking untuk latihan"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15T10:00:00Z"
                    },
                    "schedule": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/Schedule"
                            }
                        ]
                    }
                },
                "type": "object"
            },
...
            "ErrorResponse": {
                "title": "Error Response",
                "description": "Error response format",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Error occurred"
                    },
                    "error": {
                        "type": "string",
                        "example": "Detailed error message"
                    }
                },
                "type": "object"
            },
            "ValidationErrorResponse": {
                "title": "Validation Error Response",
                "description": "Validation error response format",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "properties": {
                            "field_name": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "The field is required."
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        }
    },
...
}